Migrate to Spring Boot 3.1.0

+++ server/src/main/java/com/adobe/testing/s3mock/store/ObjectStore.java
@@ -35,10 +35,10 @@ import com.adobe.testing.s3mock.util.AwsChecksumInputStream;
 import com.adobe.testing.s3mock.util.AwsChunkedDecodingChecksumInputStream;
 import com.adobe.testing.s3mock.util.AwsChunkedDecodingInputStream;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.xml.bind.JAXBException;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -51,11 +51,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import javax.xml.bind.JAXBException;
 import javax.xml.stream.XMLStreamException;
 import org.apache.commons.io.FileUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Stores objects and their metadata created in S3Mock.
@@ -65,10 +62,9 @@ public class ObjectStore {
   private static final String ACL_FILE = "objectAcl.xml";
   private static final String DATA_FILE = "binaryData";
 
-  private static final Logger LOG = LoggerFactory.getLogger(ObjectStore.class);
   /**
    * This map stores one lock object per S3Object ID.
-   * Any method modifying the underlying file must aquire the lock object before the modification.
+   * Any method modifying the underlying file must acquire the lock object before the modification.
    */
   private final Map<UUID, Object> lockStore = new ConcurrentHashMap<>();
 
@@ -115,45 +111,45 @@ public class ObjectStore {
       ChecksumAlgorithm checksumAlgorithm,
       String checksum,
       Owner owner) {
-    Instant now = Instant.now();
-    S3ObjectMetadata s3ObjectMetadata = new S3ObjectMetadata();
-    s3ObjectMetadata.setId(id);
-    s3ObjectMetadata.setKey(key);
-    s3ObjectMetadata.setContentType(contentType);
-    s3ObjectMetadata.setStoreHeaders(storeHeaders);
-    s3ObjectMetadata.setUserMetadata(userMetadata);
-    s3ObjectMetadata.setTags(tags);
-    s3ObjectMetadata.setEncryptionHeaders(encryptionHeaders);
-    s3ObjectMetadata.setModificationDate(s3ObjectDateFormat.format(now));
-    s3ObjectMetadata.setLastModified(now.toEpochMilli());
-    s3ObjectMetadata.setOwner(owner);
     lockStore.putIfAbsent(id, new Object());
     synchronized (lockStore.get(id)) {
       createObjectRootFolder(bucket, id);
-      boolean checksumEmbedded = checksumAlgorithm != null && checksum == null;
-      InputStream inputStream =
-          wrapStream(dataStream, useV4ChunkedWithSigningFormat, checksumEmbedded);
-      File dataFile =
-          inputStreamToFile(inputStream,
-              getDataFilePath(bucket, id));
-      s3ObjectMetadata.setDataPath(dataFile.toPath());
-      s3ObjectMetadata.setSize(Long.toString(dataFile.length()));
-      s3ObjectMetadata.setEtag(etag != null ? etag :
-          hexDigest(encryptionHeaders.get(X_AMZ_SERVER_SIDE_ENCRYPTION_AWS_KMS_KEY_ID), dataFile));
+      var checksumEmbedded = checksumAlgorithm != null && checksum == null;
+      var inputStream = wrapStream(dataStream, useV4ChunkedWithSigningFormat, checksumEmbedded);
+      var dataFile = inputStreamToFile(inputStream, getDataFilePath(bucket, id));
       if (inputStream instanceof AwsChecksumInputStream) {
         checksum = ((AwsChecksumInputStream) inputStream).getChecksum();
       }
-      s3ObjectMetadata.setChecksum(checksum);
-      s3ObjectMetadata.setChecksumAlgorithm(checksumAlgorithm);
-
+      var now = Instant.now();
+      var s3ObjectMetadata = new S3ObjectMetadata(
+          id,
+          key,
+          Long.toString(dataFile.length()),
+          s3ObjectDateFormat.format(now),
+          etag != null
+              ? etag
+              : hexDigest(encryptionHeaders.get(X_AMZ_SERVER_SIDE_ENCRYPTION_AWS_KMS_KEY_ID),
+                  dataFile),
+          contentType,
+          now.toEpochMilli(),
+          dataFile.toPath(),
+          userMetadata,
+          tags,
+          null,
+          null,
+          owner,
+          storeHeaders,
+          encryptionHeaders,
+          checksumAlgorithm,
+          checksum
+      );
       writeMetafile(bucket, s3ObjectMetadata);
+      return s3ObjectMetadata;
     }
-
-    return s3ObjectMetadata;
   }
 
   private AccessControlPolicy privateCannedAcl(Owner owner) {
-    Grant grant = new Grant(Grantee.from(owner), Grant.Permission.FULL_CONTROL);
+    var grant = new Grant(Grantee.from(owner), Grant.Permission.FULL_CONTROL);
     return new AccessControlPolicy(owner, Collections.singletonList(grant));
   }
 
@@ -162,13 +158,30 @@ public class ObjectStore {
    *
    * @param bucket Bucket the object is stored in.
    * @param id object ID to store tags for.
-   * @param tags List of tag objects.
+   * @param tags List of tagSet objects.
    */
   public void storeObjectTags(BucketMetadata bucket, UUID id, List<Tag> tags) {
     synchronized (lockStore.get(id)) {
-      S3ObjectMetadata s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
-      s3ObjectMetadata.setTags(tags);
-      writeMetafile(bucket, s3ObjectMetadata);
+      var s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
+      writeMetafile(bucket, new S3ObjectMetadata(
+          s3ObjectMetadata.id(),
+          s3ObjectMetadata.key(),
+          s3ObjectMetadata.size(),
+          s3ObjectMetadata.modificationDate(),
+          s3ObjectMetadata.etag(),
+          s3ObjectMetadata.contentType(),
+          s3ObjectMetadata.lastModified(),
+          s3ObjectMetadata.dataPath(),
+          s3ObjectMetadata.userMetadata(),
+          tags,
+          s3ObjectMetadata.legalHold(),
+          s3ObjectMetadata.retention(),
+          s3ObjectMetadata.owner(),
+          s3ObjectMetadata.storeHeaders(),
+          s3ObjectMetadata.encryptionHeaders(),
+          s3ObjectMetadata.checksumAlgorithm(),
+          s3ObjectMetadata.checksum()
+      ));
     }
   }
 
@@ -181,9 +194,26 @@ public class ObjectStore {
    */
   public void storeLegalHold(BucketMetadata bucket, UUID id, LegalHold legalHold) {
     synchronized (lockStore.get(id)) {
-      S3ObjectMetadata s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
-      s3ObjectMetadata.setLegalHold(legalHold);
-      writeMetafile(bucket, s3ObjectMetadata);
+      var s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
+      writeMetafile(bucket, new S3ObjectMetadata(
+          s3ObjectMetadata.id(),
+          s3ObjectMetadata.key(),
+          s3ObjectMetadata.size(),
+          s3ObjectMetadata.modificationDate(),
+          s3ObjectMetadata.etag(),
+          s3ObjectMetadata.contentType(),
+          s3ObjectMetadata.lastModified(),
+          s3ObjectMetadata.dataPath(),
+          s3ObjectMetadata.userMetadata(),
+          s3ObjectMetadata.tags(),
+          legalHold,
+          s3ObjectMetadata.retention(),
+          s3ObjectMetadata.owner(),
+          s3ObjectMetadata.storeHeaders(),
+          s3ObjectMetadata.encryptionHeaders(),
+          s3ObjectMetadata.checksumAlgorithm(),
+          s3ObjectMetadata.checksum()
+      ));
     }
   }
 
@@ -199,10 +229,10 @@ public class ObjectStore {
   }
 
   public AccessControlPolicy readAcl(BucketMetadata bucket, UUID id) {
-    AccessControlPolicy policy = readAclFile(bucket, id);
+    var policy = readAclFile(bucket, id);
     if (policy == null) {
-      S3ObjectMetadata s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
-      return privateCannedAcl(s3ObjectMetadata.getOwner());
+      var s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
+      return privateCannedAcl(s3ObjectMetadata.owner());
     }
     return policy;
   }
@@ -216,9 +246,26 @@ public class ObjectStore {
    */
   public void storeRetention(BucketMetadata bucket, UUID id, Retention retention) {
     synchronized (lockStore.get(id)) {
-      S3ObjectMetadata s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
-      s3ObjectMetadata.setRetention(retention);
-      writeMetafile(bucket, s3ObjectMetadata);
+      var s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
+      writeMetafile(bucket, new S3ObjectMetadata(
+          s3ObjectMetadata.id(),
+          s3ObjectMetadata.key(),
+          s3ObjectMetadata.size(),
+          s3ObjectMetadata.modificationDate(),
+          s3ObjectMetadata.etag(),
+          s3ObjectMetadata.contentType(),
+          s3ObjectMetadata.lastModified(),
+          s3ObjectMetadata.dataPath(),
+          s3ObjectMetadata.userMetadata(),
+          s3ObjectMetadata.tags(),
+          s3ObjectMetadata.legalHold(),
+          retention,
+          s3ObjectMetadata.owner(),
+          s3ObjectMetadata.storeHeaders(),
+          s3ObjectMetadata.encryptionHeaders(),
+          s3ObjectMetadata.checksumAlgorithm(),
+          s3ObjectMetadata.checksum()
+      ));
     }
   }
 
@@ -231,20 +278,18 @@ public class ObjectStore {
    * @return S3ObjectMetadata or null if not found
    */
   public S3ObjectMetadata getS3ObjectMetadata(BucketMetadata bucket, UUID id) {
-    S3ObjectMetadata theObject = null;
-
-    Path metaPath = getMetaFilePath(bucket, id);
+    var metaPath = getMetaFilePath(bucket, id);
 
     if (Files.exists(metaPath)) {
       synchronized (lockStore.get(id)) {
         try {
-          theObject = objectMapper.readValue(metaPath.toFile(), S3ObjectMetadata.class);
+          return objectMapper.readValue(metaPath.toFile(), S3ObjectMetadata.class);
         } catch (IOException e) {
           throw new IllegalArgumentException("Could not read object metadata-file " + id, e);
         }
       }
     }
-    return theObject;
+    return null;
   }
 
   /**
@@ -266,35 +311,32 @@ public class ObjectStore {
       String destinationKey,
       Map<String, String> encryptionHeaders,
       Map<String, String> userMetadata) {
-    S3ObjectMetadata sourceObject = getS3ObjectMetadata(sourceBucket, sourceId);
+    var sourceObject = getS3ObjectMetadata(sourceBucket, sourceId);
     if (sourceObject == null) {
       return null;
     }
-    S3ObjectMetadata copiedObject;
     synchronized (lockStore.get(sourceId)) {
-      try (InputStream inputStream = Files.newInputStream(sourceObject.getDataPath())) {
-        copiedObject = storeS3ObjectMetadata(destinationBucket,
+      try (var inputStream = Files.newInputStream(sourceObject.dataPath())) {
+        var copiedObject = storeS3ObjectMetadata(destinationBucket,
             destinationId,
             destinationKey,
-            sourceObject.getContentType(),
-            sourceObject.getStoreHeaders(),
+            sourceObject.contentType(),
+            sourceObject.storeHeaders(),
             inputStream,
             false,
             userMetadata == null || userMetadata.isEmpty()
-                ? sourceObject.getUserMetadata() : userMetadata,
+                ? sourceObject.userMetadata() : userMetadata,
             encryptionHeaders,
             null,
-            sourceObject.getTags(),
-            sourceObject.getChecksumAlgorithm(),
-            sourceObject.getChecksum(),
-            sourceObject.getOwner());
+            sourceObject.tags(),
+            sourceObject.checksumAlgorithm(),
+            sourceObject.checksum(),
+            sourceObject.owner());
+        return new CopyObjectResult(copiedObject.modificationDate(), copiedObject.etag());
       } catch (IOException e) {
-        LOG.error("Can't write file to disk!", e);
         throw new IllegalStateException("Can't write file to disk!", e);
       }
     }
-
-    return new CopyObjectResult(copiedObject.getModificationDate(), copiedObject.getEtag());
   }
 
   /**
@@ -305,16 +347,32 @@ public class ObjectStore {
   public CopyObjectResult pretendToCopyS3Object(BucketMetadata sourceBucket,
       UUID sourceId,
       Map<String, String> userMetadata) {
-    S3ObjectMetadata sourceObject = getS3ObjectMetadata(sourceBucket, sourceId);
+    var sourceObject = getS3ObjectMetadata(sourceBucket, sourceId);
     if (sourceObject == null) {
       return null;
     }
 
-    sourceObject.setLastModified(Instant.now().toEpochMilli());
-    sourceObject.setUserMetadata(userMetadata == null || userMetadata.isEmpty()
-        ? sourceObject.getUserMetadata() : userMetadata);
-    writeMetafile(sourceBucket, sourceObject);
-    return new CopyObjectResult(sourceObject.getModificationDate(), sourceObject.getEtag());
+    writeMetafile(sourceBucket, new S3ObjectMetadata(
+        sourceObject.id(),
+        sourceObject.key(),
+        sourceObject.size(),
+        sourceObject.modificationDate(),
+        sourceObject.etag(),
+        sourceObject.contentType(),
+        Instant.now().toEpochMilli(),
+        sourceObject.dataPath(),
+        userMetadata == null || userMetadata.isEmpty()
+            ? sourceObject.userMetadata() : userMetadata,
+        sourceObject.tags(),
+        sourceObject.legalHold(),
+        sourceObject.retention(),
+        sourceObject.owner(),
+        sourceObject.storeHeaders(),
+        sourceObject.encryptionHeaders(),
+        sourceObject.checksumAlgorithm(),
+        sourceObject.checksum()
+    ));
+    return new CopyObjectResult(sourceObject.modificationDate(), sourceObject.etag());
   }
 
   /**
@@ -326,13 +384,12 @@ public class ObjectStore {
    * @return true if deletion succeeded.
    */
   public boolean deleteObject(BucketMetadata bucket, UUID id) {
-    S3ObjectMetadata s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
+    var s3ObjectMetadata = getS3ObjectMetadata(bucket, id);
     if (s3ObjectMetadata != null) {
       synchronized (lockStore.get(id)) {
         try {
           FileUtils.deleteDirectory(getObjectFolderPath(bucket, id).toFile());
         } catch (IOException e) {
-          LOG.error("Can't delete directory.", e);
           throw new IllegalStateException("Can't delete directory.", e);
         }
         lockStore.remove(id);
@@ -344,7 +401,7 @@ public class ObjectStore {
   }
 
   void loadObjects(BucketMetadata bucketMetadata, Collection<UUID> ids) {
-    for (UUID id : ids) {
+    for (var id : ids) {
       lockStore.putIfAbsent(id, new Object());
       getS3ObjectMetadata(bucketMetadata, id);
     }
@@ -360,16 +417,14 @@ public class ObjectStore {
    * @return the newly created File.
    */
   File inputStreamToFile(InputStream inputStream, Path filePath) {
-    File targetFile = filePath.toFile();
+    var targetFile = filePath.toFile();
     try {
-      if (targetFile.createNewFile()) {
-        if (!retainFilesOnExit) {
-          targetFile.deleteOnExit();
-        }
+      if (targetFile.createNewFile() && (!retainFilesOnExit)) {
+        targetFile.deleteOnExit();
       }
 
-      try (InputStream is = inputStream;
-          OutputStream os = newOutputStream(targetFile.toPath())) {
+      try (var is = inputStream;
+          var os = newOutputStream(targetFile.toPath())) {
         int read;
         byte[] bytes = new byte[1024];
 
@@ -378,7 +433,6 @@ public class ObjectStore {
         }
       }
     } catch (IOException e) {
-      LOG.error("Can't write file to disk!", e);
       throw new IllegalStateException("Can't write file to disk!", e);
     }
     return targetFile;
@@ -404,13 +458,13 @@ public class ObjectStore {
    *
    * @return The Folder to store the Object in.
    */
-  private boolean createObjectRootFolder(BucketMetadata bucket, UUID id) {
-    File objectRootFolder = getObjectFolderPath(bucket, id).toFile();
-    return objectRootFolder.mkdirs();
+  private void createObjectRootFolder(BucketMetadata bucket, UUID id) {
+    var objectRootFolder = getObjectFolderPath(bucket, id).toFile();
+    objectRootFolder.mkdirs();
   }
 
   private Path getObjectFolderPath(BucketMetadata bucket, UUID id) {
-    return Paths.get(bucket.getPath().toString(), id.toString());
+    return Paths.get(bucket.path().toString(), id.toString());
   }
 
   private Path getMetaFilePath(BucketMetadata bucket, UUID id) {
@@ -426,18 +480,16 @@ public class ObjectStore {
     return Paths.get(getObjectFolderPath(bucket, id).toString(), DATA_FILE);
   }
 
-  private boolean writeMetafile(BucketMetadata bucket, S3ObjectMetadata s3ObjectMetadata) {
+  private void writeMetafile(BucketMetadata bucket, S3ObjectMetadata s3ObjectMetadata) {
     try {
-      synchronized (lockStore.get(s3ObjectMetadata.getId())) {
-        File metaFile = getMetaFilePath(bucket, s3ObjectMetadata.getId()).toFile();
+      synchronized (lockStore.get(s3ObjectMetadata.id())) {
+        var metaFile = getMetaFilePath(bucket, s3ObjectMetadata.id()).toFile();
         if (!retainFilesOnExit) {
           metaFile.deleteOnExit();
         }
         objectMapper.writeValue(metaFile, s3ObjectMetadata);
-        return true;
       }
     } catch (IOException e) {
-      LOG.error("Could not write object metadata-file.", e);
       throw new IllegalStateException("Could not write object metadata-file.", e);
     }
   }
@@ -445,31 +497,28 @@ public class ObjectStore {
   private AccessControlPolicy readAclFile(BucketMetadata bucket, UUID id) {
     try {
       synchronized (lockStore.get(id)) {
-        File aclFile = getAclFilePath(bucket, id).toFile();
+        var aclFile = getAclFilePath(bucket, id).toFile();
         if (!aclFile.exists()) {
           return null;
         }
-        String toDeserialize = FileUtils.readFileToString(aclFile, Charset.defaultCharset());
+        var toDeserialize = FileUtils.readFileToString(aclFile, Charset.defaultCharset());
         return deserializeJaxb(toDeserialize);
       }
     } catch (IOException | JAXBException | XMLStreamException e) {
-      LOG.error("Could not write object metadata-file.", e);
       throw new IllegalStateException("Could not write object metadata-file.", e);
     }
   }
 
-  private boolean writeAclFile(BucketMetadata bucket, UUID id, AccessControlPolicy policy) {
+  private void writeAclFile(BucketMetadata bucket, UUID id, AccessControlPolicy policy) {
     try {
       synchronized (lockStore.get(id)) {
-        File aclFile = getAclFilePath(bucket, id).toFile();
+        var aclFile = getAclFilePath(bucket, id).toFile();
         if (!retainFilesOnExit) {
           aclFile.deleteOnExit();
         }
         FileUtils.write(aclFile, serializeJaxb(policy), Charset.defaultCharset());
-        return true;
       }
     } catch (IOException | JAXBException e) {
-      LOG.error("Could not write object metadata-file.", e);
       throw new IllegalStateException("Could not write object metadata-file.", e);
     }
   }
